+++ packages/core/src/__tests__/data/types/type-external-store.raw.tsx
@@ -1,0 +1,10 @@
+import { useStore } from '@builder.io/mitosis';
+import { FooStore } from './foo-store';
+
+export default function TypeExternalStore() {
+  const state = useStore<FooStore>({
+    _name: 'test',
+  });
+
+  return <div>Hello {state._name}! </div>;
+}