+++ packages/core/src/generators/angular/index.ts
@@ -1,6 +1,7 @@
 import { SELF_CLOSING_HTML_TAGS, VALID_HTML_TAGS } from '@/constants/html_tags';
 import { dedent } from '@/helpers/dedent';
 import { fastClone } from '@/helpers/fast-clone';
+import { getChildComponents } from '@/helpers/get-child-components';
 import { getComponentsUsed } from '@/helpers/get-components-used';
 import { getCustomImports } from '@/helpers/get-custom-imports';
 import { getPropFunctions } from '@/helpers/get-prop-functions';
@@ -8,6 +9,7 @@ import { getProps } from '@/helpers/get-props';
 import { getPropsRef } from '@/helpers/get-props-ref';
 import { getRefs } from '@/helpers/get-refs';
 import { getStateObjectStringFromComponent } from '@/helpers/get-state-object-string';
+import { getTypedFunction } from '@/helpers/get-typed-function';
 import { indent } from '@/helpers/indent';
 import { isMitosisNode } from '@/helpers/is-mitosis-node';
 import { isUpperCase } from '@/helpers/is-upper-case';
@@ -47,7 +49,9 @@ import {
   addCodeToOnInit,
   addCodeToOnUpdate,
   getAppropriateTemplateFunctionKeys,
+  getDefaultProps,
   makeReactiveState,
+  transformState,
 } from './helpers';
 import {
   AngularBlockOptions,
@@ -220,7 +224,8 @@ const stringifyBinding =
 
     if (keyToUse.startsWith('on')) {
       const { event, value } = processEventBinding(keyToUse, code, node.name, cusArgs[0]);
-      return ` (${event})="${value}"`;
+      // Angular events are all lowerCased
+      return ` (${event.toLowerCase()})="${value}"`;
     } else if (keyToUse === 'class') {
       return ` [class]="${code}" `;
     } else if (keyToUse === 'ref' || keyToUse === 'spreadRef') {
@@ -722,25 +727,6 @@ const classPropertiesPlugin = () => ({
   },
 });
 
-// if any state "property" is trying to access state.* or props.*
-// then we need to move them to onInit where they can be accessed
-const transformState = (json: MitosisComponent) => {
-  Object.entries(json.state)
-    .reverse()
-    .forEach(([key, value]) => {
-      if (value?.type === 'property') {
-        if (value.code && (value.code.includes('state.') || value.code.includes('props.'))) {
-          const code = stripStateAndPropsRefs(value.code, { replaceWith: 'this' });
-          json.state[key]!.code = 'null';
-          if (!json.hooks.onInit?.code) {
-            json.hooks.onInit = { code: '' };
-          }
-          json.hooks.onInit.code = `\nthis.${key} = ${code};\n${json.hooks.onInit.code}`;
-        }
-      }
-    });
-};
-
 export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
   (userOptions = {}) =>
   ({ component: _component }) => {
@@ -824,16 +810,9 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
     }
 
     const [forwardProp, hasPropRef] = getPropsRef(json, true);
-    const childComponents: string[] = [json.name]; // a component can be recursively used in itself
     const propsTypeRef = json.propsTypeRef !== 'any' ? json.propsTypeRef : undefined;
 
-    json.imports.forEach(({ imports }) => {
-      Object.keys(imports).forEach((key) => {
-        if (imports[key] === 'default') {
-          childComponents.push(key);
-        }
-      });
-    });
+    const childComponents: string[] = getChildComponents(json);
 
     const customImports = getCustomImports(json);
 
@@ -934,13 +913,21 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
 
     const dataString = getStateObjectStringFromComponent(json, {
       format: 'class',
-      valueMapper: processAngularCode({
-        replaceWith: 'this',
-        contextVars,
-        outputVars,
-        domRefs: Array.from(domRefs),
-        stateVars,
-      }),
+      withType: options.typescript,
+      valueMapper: (code, type, typeParameter) => {
+        let value = code;
+        if (type !== 'data') {
+          value = getTypedFunction(code, options.typescript, typeParameter);
+        }
+
+        return processAngularCode({
+          replaceWith: 'this',
+          contextVars,
+          outputVars,
+          domRefs: Array.from(domRefs),
+          stateVars,
+        })(value);
+      },
     });
 
     const refsForObjSpread = getRefs(json, 'spreadRef');
@@ -975,19 +962,6 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
       componentMetadata[key] = value;
     });
 
-    const getPropsDefinition = ({ json }: { json: MitosisComponent }) => {
-      if (!json.defaultProps) return '';
-      const defalutPropsString = Object.keys(json.defaultProps)
-        .map((prop) => {
-          const value = json.defaultProps!.hasOwnProperty(prop)
-            ? json.defaultProps![prop]?.code
-            : 'undefined';
-          return `${prop}: ${value}`;
-        })
-        .join(',');
-      return `const defaultProps = {${defalutPropsString}};\n`;
-    };
-
     const hasConstructor =
       Boolean(injectables.length) || dynamicComponents.size || refsForObjSpread.size;
 
@@ -1009,7 +983,7 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
     ${options.standalone ? `import { CommonModule } from '@angular/common';` : ''}
 
     ${json.types ? json.types.join('\n') : ''}
-    ${getPropsDefinition({ json })}
+    ${getDefaultProps(json)}
     ${renderPreComponent({
       explicitImportFileExtension: options.explicitImportFileExtension,
       component: json,