+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @builder.io/mitosis-cli
 
+## 0.5.0
+
+### Patch Changes
+
+- Updated dependencies [7e2c95f]
+- Updated dependencies [4171a19]
+- Updated dependencies [d59d328]
+  - @builder.io/mitosis@0.5.0
+
+## 0.4.7
+
+### Patch Changes
+
+- Updated dependencies [068efab]
+  - @builder.io/mitosis@0.4.7
+
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies [a7fd87f]
+- Updated dependencies [a7fd87f]
+  - @builder.io/mitosis@0.4.6
+
 ## 0.4.5
 
 ### Patch Changes